requirements.txt
fastapi
bcrypt
pyjwt
alembic
uvicorn
SQLAlchemy
psycopg2-binary
email-validator
python-dotenv
python-jose